make writev and readv declarations match susv3

+++ include/sys/uio.h
@@ -3,7 +3,8 @@
 
 #include <sys/socket.h>
 
-int readv(int filedes, const struct iovec *vector, size_t count);
-int writev(int filedes, const struct iovec *vector, size_t count);
+/* I have no idea why susv3 specifies count as int instead of size_t */
+ssize_t readv(int filedes, const struct iovec *vector, int count);
+ssize_t writev(int filedes, const struct iovec *vector, int count);
 
 #endif
